// data/github/current/ui/packages/issues-react/components/search/__tests__/__generated__/SearchTestQuery.graphql.ts
export type SearchTestQuery$variables = {
  first?: number | null | undefined;
  labelPageSize?: number | null | undefined;
  query?: string | null | undefined;
  skip?: number | null | undefined;
};
export type SearchTestQuery$data = {
  readonly " $fragmentSpreads": FragmentRefs<"SearchRootFragment">;
};
export type SearchTestQuery = {
  response: SearchTestQuery$data;
  variables: SearchTestQuery$variables;
};